simplify test commands
