+++ .github/workflows/test_and_deploy.yml
@@ -40,11 +40,13 @@ jobs:
       - name: Setup test database
         run: bash ./testing/configure-db.sh
 
-      - name: Start test server
-        run: npm run start:test:ci &
-
       - name: Run tests
-        run: npm run test:server:ci
+        run: npm run test:ci
+
+      - name: Run server integration tests
+        run: |
+          npm run start:test:ci &
+          npm run test:server:ci
 
   deploy:
     name: Deploy to production
