+++ jest.config-browser.ts
@@ -1,0 +1,41 @@
+import createJestConfig from 'testing/createJestConfig';
+/*
+ * For a detailed explanation regarding each configuration property and type check, visit:
+ * https://jestjs.io/docs/configuration
+ */
+
+/**
+ * Configuration copied from
+ * https://nextjs.org/docs/testing
+ * */
+export const jestConfig = {
+
+  // Automatically clear mock calls, instances and results before every test
+  clearMocks: true,
+
+  // Indicates whether the coverage information should be collected while executing the test
+  collectCoverage: false,
+
+  // A path to a module which exports an async function that is triggered once before all test suites
+  globalSetup: '<rootDir>/testing/setupDatabase.ts',
+
+  // An array of directory names to be searched recursively up from the requiring module's location
+  moduleDirectories: [
+    'node_modules',
+    '<rootDir>'
+  ],
+
+  // The test environment that will be used for testing
+  testEnvironment: 'jsdom',
+
+  // The glob patterns Jest uses to detect test files
+  testMatch: [
+    '**/components/**/?(*.)+(spec).[tj]s?(x)',
+    '**/components/**/**.test.tsx'
+  ],
+
+  testTimeout: 30000
+
+};
+
+export default createJestConfig(jestConfig);
