+++ package.json
@@ -12,7 +12,7 @@
     "cron:staging": "npx react-env --path .env.staging -- node ./.next/server/cron.js",
     "cron:prod": "dotenv -e .env.production -- node ./.next/server/cron.js",
     "cron:dev": "dotenv -e .env.local -- ts-node ./background/cron.ts",
-    "migrate:test": "dotenv -e .env.test.local -- npx prisma migrate dev",
+    "migrate:test": "dotenv -e .env.test.local -- npx prisma migrate deploy",
     "migrate:prod": "npx dotenv -e .env.staging -- prisma migrate deploy",
     "migrate:staging": "npx dotenv -e .env.staging -- prisma migrate deploy",
     "server:dev": "next dev",
@@ -31,12 +31,12 @@
     "db-tool:dev": "dotenv -e .env.local -- npx prisma studio",
     "db-tool:test": "dotenv -e .env.test.local -- npx prisma studio",
     "test:setup-db": "./testing/configure-db.sh",
-    "test:client": "npm run test-setup && jest --config='./jest.config-client.ts'",
-    "test:migrate": "dotenv -e .env.local -- npx prisma migrate dev",
-    "test:server": "dotenv -e .env.test.local --override -- npx jest --config='./jest.config-server.ts' --verbose --watch",
-    "test:server:ci": "dotenv -e .env.test.local --override -- npx jest --config='./jest.config-server.ts'",
-    "test:server:debug": "dotenv -e .env.test.local --override -- node --inspect-brk --expose-gc ./node_modules/.bin/jest --runInBand --logHeapUsage --config='./jest.config-server.ts' --verbose",
-    "test:server-integration": "dotenv -e .env.test.local --override -- npx jest --config='./jest.config-server-integration.ts' --verbose"
+    "test:browser": "npm run test-setup && jest --config='./jest.config-client.ts'",
+    "test": "npm run test:ci -- --verbose --watch",
+    "test:ci": "dotenv -e .env.test.local --override -- npx jest --config='./jest.config.ts'",
+    "test:debug": "dotenv -e .env.test.local --override -- node --inspect-brk --expose-gc ./node_modules/.bin/jest --runInBand --logHeapUsage --config='./jest.config-server.ts' --verbose",
+    "test:server": "npm run test:server:ci -- --verbose --watch",
+    "test:server:ci": "dotenv -e .env.test.local --override -- npx jest --config='./jest.config-server.ts'"
   },
   "prepare": "husky install",
   "lint-staged": {
